# target/classes/api.yml
openapi: 3.0.3
info:
  title: Credit Service API
  version: 4.0.0
  description: >
    Credit products API.
    - Loans (PERSONAL/BUSINESS) live under /credits
    - Credit card lines live under /credit-lines
servers:
  - url: http://localhost:8585
...
paths:
...
  /credit-products/{id}:
    get:
      tags: [ Common ]
      summary: Get credit product (loan or credit line)
      operationId: getCreditProduct
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Credit product summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditProductSummary' }
        '404': { $ref: '#/components/responses/NotFound' }
  /credit-products/{id}/balance:
    get:
      tags: [ Common ]
      summary: Get product balance
      operationId: getCreditProductBalance
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditBalance' }
        '404': { $ref: '#/components/responses/NotFound' }
  /credit-products/{id}/movements:
    get:
      tags: [ Common ]
      summary: List product movements
      operationId: listCreditProductMovements
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: type
          schema:
            type: string
            enum: [ CHARGE, PAYMENT, INTEREST, FEE, REVERSAL ]
        - in: query
          name: limit
          schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Movement list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CreditMovement' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    CreditStatus:
      type: string
      description: Estado del crédito.
      enum: [ ACTIVE, CLOSED, DELINQUENT ]
...
    CreditBalance:
      type: object
      description: Product balances (limit, used, available). # Saldos
      required: [ limit, balance, available ]
      additionalProperties: false
      properties:
        limit: { type: number, minimum: 0 }
        balance: { type: number, minimum: 0 }
        available: { type: number }
    CreditMovement:
      type: object
      description: Movement (charge, payment, interest, fee or reversal). # Movimiento
      required: [ id, creditId, type, amount, txnAt ]
      additionalProperties: false
      properties:
        id: { type: string, description: Movement id }
        creditId: { type: string, description: Product id (loan or credit line) } # Id del producto
        type:
          type: string
          enum: [ CHARGE, PAYMENT, INTEREST, FEE, REVERSAL ]
        amount: { type: number }
        txnAt: { type: string, format: date-time }
        postedAt: { type: string, format: date-time }
        runningBalance: { type: number }
        channel: { type: string }
...
    Error:
      type: object
      description: RFC 7807 Problem Details. # Estructura de error
      additionalProperties: false
      properties:
        type: { type: string, example: 'https://errors.bank.com/business-rule' }
        title: { type: string, example: 'Business rule violation' }
        status: { type: integer, example: 409 }
        detail: { type: string, example: 'Customer already has an active PERSONAL loan.' }
        instance: { type: string, example: '/credits/12345' }
        code: { $ref: '#/components/schemas/ErrorCode' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
    FieldError:
      type: object
      description: Field validation error. # Error por campo
      additionalProperties: false
      properties:
        field: { type: string, example: 'interestAnnual' }
        message: { type: string, example: 'Must be greater or equal than 0' }
        code: { type: string, example: 'Min' }
    ErrorCode:
      type: string
      description: Business code. # Código interno
      enum:
        - CREDIT_PERSONAL_LIMIT
        - CREDIT_DELINQUENT_LOCK
        - CREDIT_CARD_LIMIT_EXCEEDED
        - CREDIT_INVALID_STATE
        - CREDIT_NOT_FOUND
        - CUSTOMER_NOT_FOUND
        - VALIDATION_ERROR
        - SERVICE_UNAVAILABLE

  # Credit Lines
...
    CreditProductSummary:
      type: object
      description: Unified view for any credit product (loan or credit line).
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [ LOAN, CREDIT_LINE ]
        status: { $ref: '#/components/schemas/CreditStatus' }
        limit: { type: number }
        balance: { type: number }
        currency: { type: string, example: PEN }
        createdAt: { type: string, format: date-time }